# brew.sh: speed up `brew --help`, `brew -h`, etc.